+++ test/recordChildProcess.js
@@ -1,37 +1,42 @@
 'use strict';
 
 /**
+ * A child process recording.
+ * @kind typedef
+ * @name ChildProcessRecording
+ * @type {object}
+ * @prop {string} stdout The child process `stdout`.
+ * @prop {string} stderr The child process `stderr`.
+ * @prop {number} [exitCode] The child process exit code if it exited on its own.
+ * @prop {string} [signal] The signal by which the child process was terminated if it didn’t exit on its own.
+ * @ignore
+ */
+
+/**
  * Records a Node.js child process.
  * @kind function
  * @name recordChildProcess
  * @param {ChildProcess} childProcess Node.js child process.
- * @returns {Promise<{output: {stdout: string, stderr: string}, exitCode: number, signal: string}>} Resolves a multidimensional array of output types (`stderr` or `stdout`) and values, the exit code if the child exited on its own, or the signal by which the child process was terminated.
+ * @returns {Promise<ChildProcessRecording>} Resolves the recording of the child process.
  * @ignore
  */
 module.exports = function recordChildProcess(childProcess) {
   return new Promise((resolve, reject) => {
-    const output = {
-      stdout: '',
-      stderr: '',
-    };
-
-    const add = (pipe, data) => {
-      const text = data.toString();
-      output[pipe] += text;
-    };
+    let stdout = '';
+    let stderr = '';
 
     childProcess.stdout.on('data', (data) => {
-      add('stdout', data);
+      stdout += data.toString();
     });
 
     childProcess.stderr.on('data', (data) => {
-      add('stderr', data);
+      stderr += data.toString();
     });
 
     childProcess.once('error', reject);
 
     childProcess.once('close', (exitCode, signal) =>
-      resolve({ output, exitCode, signal })
+      resolve({ stdout, stderr, exitCode, signal })
     );
   });
 };